requirement.txt
fastapi
uvicorn
nest-asyncio
torch
Pillow
torchvision
python-dotenv
